Alert dialog for logout

+++ app/src/main/java/com/solusianakbangsa/gameyourfit/HomeActivity.kt
@@ -1,15 +1,16 @@
 package com.solusianakbangsa.gameyourfit
 
+import android.content.DialogInterface
 import android.content.Intent
+import android.content.SharedPreferences
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
-import android.util.Log
 import android.view.Menu
+import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
-import com.google.android.material.floatingactionbutton.FloatingActionButton
-import com.google.android.material.snackbar.Snackbar
+import androidx.appcompat.app.AlertDialog
 import com.google.android.material.navigation.NavigationView
 import androidx.navigation.findNavController
 import androidx.navigation.ui.AppBarConfiguration
@@ -21,16 +22,17 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.appcompat.widget.Toolbar
 import androidx.preference.PreferenceManager
 import com.firebase.ui.auth.AuthUI
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.FirebaseDatabase
 import com.solusianakbangsa.gameyourfit.ui.ImageReplacer
 import com.solusianakbangsa.gameyourfit.ui.auth.LoginActivity
 import kotlinx.android.synthetic.main.nav_header_main.*
 import java.io.File
-import java.net.URL
 
 class HomeActivity : AppCompatActivity() {
 
+    lateinit var sharedPref: SharedPreferences
     private lateinit var appBarConfiguration: AppBarConfiguration
     private val imageReplacer = ImageReplacer()
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -43,23 +45,27 @@ class HomeActivity : AppCompatActivity() {
         val navView: NavigationView = findViewById(R.id.nav_view)
         val navController = findNavController(R.id.nav_host_fragment)
         val handler = Handler(Looper.getMainLooper())
-        val drawerProfilePicture = navView.getHeaderView(0).findViewById<ImageView>(R.id.drawerProfilePicture)
+        val drawerProfilePicture =
+            navView.getHeaderView(0).findViewById<ImageView>(R.id.drawerProfilePicture)
 
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
-        appBarConfiguration = AppBarConfiguration(setOf(
-            R.id.nav_dashboard, R.id.nav_leaderboard, R.id.nav_friends,R.id.nav_setting), drawerLayout)
+        appBarConfiguration = AppBarConfiguration(
+            setOf(
+                R.id.nav_dashboard, R.id.nav_leaderboard, R.id.nav_friends, R.id.nav_setting
+            ), drawerLayout
+        )
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
 
         val userId = FirebaseAuth.getInstance().uid.toString()
+        sharedPref = PreferenceManager.getDefaultSharedPreferences(this)
         val ref = FirebaseDatabase.getInstance().getReference("users").child(userId)
-        val sharedPref = PreferenceManager.getDefaultSharedPreferences(this)
-        var name : String = "Username"
+        var name: String = "Username"
 
-        if(sharedPref.contains("username")){
-            name =  sharedPref.getString("username","")!!
-        } else{
+        if (sharedPref.contains("username")) {
+            name = sharedPref.getString("username", "")!!
+        } else {
             ref.child("username").get().addOnSuccessListener {
                 name = it.value.toString()
                 sharedPref.edit().putString("username", it.value.toString()).apply()
@@ -68,8 +74,8 @@ class HomeActivity : AppCompatActivity() {
         navView.getHeaderView(0).findViewById<TextView>(R.id.drawerName).text = name
 
         val profilePicture = File(this.filesDir, FileConstants.PROFILE_PICTURE_FILENAME)
-        if(!(profilePicture.exists())){
-            ref.child("images").get().addOnSuccessListener{
+        if (!(profilePicture.exists())) {
+            ref.child("images").get().addOnSuccessListener {
                 imageReplacer.replaceImage(
                     handler,
                     drawerProfilePicture,
@@ -79,38 +85,56 @@ class HomeActivity : AppCompatActivity() {
                     FileConstants.PROFILE_PICTURE_FILENAME
                 )
             }
-        } else{
-            imageReplacer.replaceImage(drawerProfilePicture, this, FileConstants.PROFILE_PICTURE_FILENAME)
+        } else {
+            imageReplacer.replaceImage(
+                drawerProfilePicture,
+                this,
+                FileConstants.PROFILE_PICTURE_FILENAME
+            )
         }
 
-        navView.getHeaderView(0).findViewById<TextView>(R.id.drawerEmail).text = FirebaseAuth.getInstance().currentUser?.email.toString()
+        navView.getHeaderView(0).findViewById<TextView>(R.id.drawerEmail).text =
+            FirebaseAuth.getInstance().currentUser?.email.toString()
 //          One time initialization for the levels.json, campaignActivity will later read from this
 
         val logout = navView.getHeaderView(0).findViewById<TextView>(R.id.logout)
         logout.setOnClickListener {
-            AuthUI.getInstance().signOut(this).addOnCompleteListener{
-                val loginIntent = Intent(this, LoginActivity::class.java)
-                startActivity(loginIntent)
-                val file = File(this.filesDir, FileConstants.PROFILE_PICTURE_FILENAME)
-                file.delete()
-                sharedPref.edit().clear().apply()
-                finish()
-            }
+            MaterialAlertDialogBuilder(this, R.style.AlertDialogTheme)
+                .setTitle("Logout")
+                .setMessage("Are you sure you want to logout of this account?")
+                .setNegativeButton("Cancel") { dialog, which ->
+                    // Respond to negative button press
+                }
+                .setPositiveButton("Logout") { dialog, which ->
+                    AuthUI.getInstance().signOut(this).addOnCompleteListener {
+                        val loginIntent = Intent(this, LoginActivity::class.java)
+                        startActivity(loginIntent)
+                        val file = File(this.filesDir, FileConstants.PROFILE_PICTURE_FILENAME)
+                        file.delete()
+                        sharedPref.edit().clear().apply()
+                        finish()
+                    }
+                }
         }
     }
 
-    override fun onCreateOptionsMenu(menu: Menu): Boolean {
+        override fun onCreateOptionsMenu(menu: Menu): Boolean {
 //      Disables option menu (the one with 3 dots)
-        return true
-    }
+            return true
+        }
 
-    override fun onResume() {
-        super.onResume()
-        imageReplacer.replaceImage(findViewById<NavigationView>(R.id.nav_view).getHeaderView(0).findViewById(R.id.drawerProfilePicture), this, FileConstants.PROFILE_PICTURE_FILENAME)
-    }
-    override fun onSupportNavigateUp(): Boolean {
-        val navController = findNavController(R.id.nav_host_fragment)
-        return navController.navigateUp(appBarConfiguration) || super.onSupportNavigateUp()
-    }
+        override fun onResume() {
+            super.onResume()
+            imageReplacer.replaceImage(
+                findViewById<NavigationView>(R.id.nav_view).getHeaderView(0)
+                    .findViewById(R.id.drawerProfilePicture),
+                this,
+                FileConstants.PROFILE_PICTURE_FILENAME
+            )
+        }
 
-}
+        override fun onSupportNavigateUp(): Boolean {
+            val navController = findNavController(R.id.nav_host_fragment)
+            return navController.navigateUp(appBarConfiguration) || super.onSupportNavigateUp()
+        }
+    }
